requirements.txt
streamlit==1.41.1
fastapi==0.115.7
langchain==0.3.15
langchain-community==0.3.15
pydantic==2.10.6
uvicorn==0.34.0
weaviate-client==4.10.4
python-dotenv==1.0.1
sentence-transformers==3.4.0
langchain-groq==0.2.3
pypdf==5.2.0